Organize models
Separate model for each package into separate file for easier
maintenance.
