+++ limbook_api/comments/routes.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, jsonify, abort, request
 
 from limbook_api.auth.auth import requires_auth
-from limbook_api.db.model import Comment
+from limbook_api.models.comment import Comment
 
 comments = Blueprint('comments', __name__)
