+++ limbook_api/config_test.py
@@ -11,7 +11,7 @@ class Config:
 
     # Configure SQLAlchemy and database
     # SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///test.db?check_same_thread=False'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///test.models?check_same_thread=False'
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     SQLALCHEMY_ECHO = False
