+++ limbook_api/tests/test_reacts.py
@@ -4,8 +4,8 @@ from flask import json
 
 from limbook_api import create_app
 from limbook_api.config_test import Config
-from limbook_api.db.model import create_random_react, db_drop_and_create_all, \
-    create_random_post
+from limbook_api.models.post import create_random_post
+from limbook_api.models.setup import db_drop_and_create_all
 
 test_user_id = "auth0|5eb66a2d1cc1ac0c1496c16f"
